move the LargeAssignments lint logic into its own file
